OD20DEVAT3-21 Corrections from PR #1 and core EAV implementation. Changing some classes to match changes.

+++ src/main/java/com/netcracker/odstc/logviewer/models/eaventity/mappers/ObjectMapper.java
@@ -2,30 +2,23 @@ package com.netcracker.odstc.logviewer.models.eaventity.mappers;
 
 import org.springframework.jdbc.core.RowMapper;
 
-import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * Description:
- *
- * @author Aleksanid
- * created 04.12.2020
- */
 public class ObjectMapper implements RowMapper<List<Object>> {
     @Override
     public List<Object> mapRow(ResultSet resultSet, int i) throws SQLException {
         List<Object> objectColumns = new ArrayList<>();
 
         Object parent = resultSet.getObject("PARENT_ID");
-        if(parent==null){
+        if (parent == null) {
             objectColumns.add(null);
-        }else{
-            objectColumns.add(BigInteger.valueOf(resultSet.getLong("PARENT_ID")));
+        } else {
+            objectColumns.add(resultSet.getBigDecimal("PARENT_ID").toBigInteger());
         }
-        objectColumns.add(BigInteger.valueOf(resultSet.getLong("OBJECT_TYPE_ID")));
+        objectColumns.add(resultSet.getBigDecimal("OBJECT_TYPE_ID").toBigInteger());
         objectColumns.add(resultSet.getString("NAME"));
 
         return objectColumns;
